requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
dulwich==0.21.6
fastapi==0.103.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httptools==0.6.0
Hypercorn==0.15.0
hyperframe==6.0.1
idna==3.4
minio==7.1.17
motor==3.3.1
orjson==3.9.9
priority==2.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pymongo==4.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
wsproto==1.2.0
